Testing shapes and CSP psatz
Both are a fair waste of time.

The CSP psatz does not have a proper chordal completion implementation, I need to figure out the elimination trees.

The rotating shape problem doubles the order, and the LMIs are still overly large. The constraint w(A(s; omega)) >= z(omega) over S times Omega does not contain correlative sparsity. It is `unblockable' under term sparsity, but may be decomposed under the `clique' setting at the cost of a lower quality objective.